initial retructuring for cleaning up and extracting different portions out of [Default]CloudSdkService
