+++ google-cloud-tools-plugin/src/com/google/cloud/tools/intellij/appengine/sdk/DefaultCloudSdkService.java
@@ -17,34 +17,12 @@
 package com.google.cloud.tools.intellij.appengine.sdk;
 
 import com.google.cloud.tools.appengine.api.AppEngineException;
-import com.google.cloud.tools.appengine.api.whitelist.AppEngineJreWhitelist;
-import com.google.cloud.tools.appengine.cloudsdk.AppEngineJavaComponentsNotInstalledException;
 import com.google.cloud.tools.appengine.cloudsdk.CloudSdk;
-import com.google.cloud.tools.appengine.cloudsdk.CloudSdkNotFoundException;
-import com.google.cloud.tools.appengine.cloudsdk.CloudSdkOutOfDateException;
 import com.google.cloud.tools.intellij.analytics.UsageTrackerProvider;
 import com.google.cloud.tools.intellij.util.GctTracking;
-import com.google.common.annotations.VisibleForTesting;
-import com.intellij.execution.configurations.ParametersList;
 import com.intellij.ide.util.PropertiesComponent;
-import com.intellij.openapi.diagnostic.Logger;
-import com.intellij.openapi.util.text.StringUtil;
-import gnu.trove.THashMap;
-import gnu.trove.THashSet;
-import java.io.BufferedReader;
-import java.io.File;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.nio.charset.StandardCharsets;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.util.ArrayList;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
-import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 /**
@@ -52,20 +30,10 @@ import org.jetbrains.annotations.Nullable;
  * serialization.
  */
 // TODO (eshaul) Offload path logic for retrieving AE libs to the common library once implemented
-public class DefaultCloudSdkService extends CloudSdkService {
-
-  private static final Logger logger = Logger.getInstance(DefaultCloudSdkService.class);
+public class DefaultCloudSdkService implements CloudSdkService {
 
   private PropertiesComponent propertiesComponent = PropertiesComponent.getInstance();
   private static final String CLOUD_SDK_PROPERTY_KEY = "GCT_CLOUD_SDK_HOME_PATH";
-  private static final Path JAVA_TOOLS_RELATIVE_PATH =
-      Paths.get("platform", "google_appengine", "google", "appengine", "tools", "java");
-
-  // Kept around for AppEngineGwtServer
-  public static final Path LIB_APPENGINE_TOOLS_API_JAR =
-      Paths.get("lib", "appengine-tools-api.jar");
-
-  private Map<String, Set<String>> myMethodsBlackList;
 
   @Nullable
   @Override
@@ -91,152 +59,5 @@ public class DefaultCloudSdkService extends CloudSdkService {
   @Override
   public void setSdkHomePath(String cloudSdkHomePath) {
     propertiesComponent.setValue(CLOUD_SDK_PROPERTY_KEY, cloudSdkHomePath);
-  }
-
-  @Override
-  protected Set<CloudSdkValidationResult> validateCloudSdk(Path path) {
-    Set<CloudSdkValidationResult> validationResults = new HashSet<>();
-
-    if (path == null) {
-      validationResults.add(CloudSdkValidationResult.CLOUD_SDK_NOT_FOUND);
-      // If the Cloud SDK is not found, don't bother checking anything else
-      return validationResults;
-    }
-
-    CloudSdk sdk = buildCloudSdkWithPath(path);
-    try {
-      sdk.validateCloudSdk();
-    } catch (CloudSdkNotFoundException exception) {
-      validationResults.add(CloudSdkValidationResult.CLOUD_SDK_NOT_FOUND);
-      // If the Cloud SDK is not found, don't bother checking anything else
-      return validationResults;
-    } catch (CloudSdkOutOfDateException exception) {
-      validationResults.add(CloudSdkValidationResult.CLOUD_SDK_VERSION_NOT_SUPPORTED);
-    }
-
-    try {
-      sdk.validateAppEngineJavaComponents();
-    } catch (AppEngineJavaComponentsNotInstalledException ex) {
-      validationResults.add(CloudSdkValidationResult.NO_APP_ENGINE_COMPONENT);
-    }
-
-    return validationResults;
-  }
-
-  @Nullable
-  private Path getJavaToolsBasePath() {
-    return getSdkHomePath() != null
-        ? getSdkHomePath().resolve(JAVA_TOOLS_RELATIVE_PATH.toString())
-        : null;
-  }
-
-  @Nullable
-  @Override
-  // TODO(eshaul) used only by AppEngineGwtServer - can be removed if GWT support is removed
-  public File getToolsApiJarFile() {
-    return getJavaToolsBasePath() != null
-        ? getJavaToolsBasePath().resolve(LIB_APPENGINE_TOOLS_API_JAR).toFile()
-        : null;
-  }
-
-  @NotNull
-  @Override
-  public File[] getLibraries() {
-    return getJavaToolsBasePath() != null
-        ? getJarsFromDirectory(getJavaToolsBasePath().resolve(Paths.get("lib", "shared")).toFile())
-        : new File[0];
-  }
-
-  @NotNull
-  @Override
-  public File[] getJspLibraries() {
-    return getJavaToolsBasePath() != null
-        ? getJarsFromDirectory(
-            getJavaToolsBasePath().resolve(Paths.get("lib", "shared", "jsp")).toFile())
-        : new File[0];
-  }
-
-  @Override
-  public boolean isMethodInBlacklist(@NotNull String className, @NotNull String methodName) {
-    if (myMethodsBlackList == null) {
-      try {
-        myMethodsBlackList = loadBlackList();
-      } catch (IOException ioe) {
-        logger.error(ioe);
-        myMethodsBlackList = new THashMap<>();
-      }
-    }
-    final Set<String> methods = myMethodsBlackList.get(className);
-    return methods != null && methods.contains(methodName);
-  }
-
-  @Override
-  public boolean isClassInWhiteList(@NotNull String className) {
-    return AppEngineJreWhitelist.contains(className);
-  }
-
-  @Nullable
-  @Override
-  public File getWebSchemeFile() {
-    return getJavaToolsBasePath() != null
-        ? getJavaToolsBasePath().resolve(Paths.get("docs", "appengine-web.xsd")).toFile()
-        : null;
-  }
-
-  @Override
-  public void patchJavaParametersForDevServer(@NotNull ParametersList vmParameters) {
-    if (getJavaToolsBasePath() != null) {
-      File agentPath =
-          getJavaToolsBasePath().resolve(Paths.get("lib", "agent", "appengine-agent.jar")).toFile();
-      if (agentPath.exists()) {
-        vmParameters.add("-javaagent:" + agentPath.getAbsolutePath());
-      }
-      File patchPath =
-          getJavaToolsBasePath()
-              .resolve(Paths.get("lib", "override", "appengine-dev-jdk-overrides.jar"))
-              .toFile();
-      if (patchPath.exists()) {
-        vmParameters.add("-Xbootclasspath/p:" + patchPath.getAbsolutePath());
-      }
-    }
-  }
-
-  private Map<String, Set<String>> loadBlackList() throws IOException {
-    final InputStream stream = getClass().getResourceAsStream("/data/methodsBlacklist.txt");
-    logger.assertTrue(stream != null, "/data/methodsBlacklist.txt not found");
-    final THashMap<String, Set<String>> map = new THashMap<>();
-    BufferedReader reader =
-        new BufferedReader(new InputStreamReader(stream, StandardCharsets.UTF_8));
-    try {
-      String line;
-      while ((line = reader.readLine()) != null) {
-        final int i = line.indexOf(':');
-        String className = line.substring(0, i);
-        String methods = line.substring(i + 1);
-        map.put(className, new THashSet<>(StringUtil.split(methods, ",")));
-      }
-    } finally {
-      reader.close();
-    }
-    return map;
-  }
-
-  private File[] getJarsFromDirectory(File libFolder) {
-    List<File> jars = new ArrayList<>();
-    final File[] files = libFolder.listFiles();
-    if (files != null) {
-      for (File file : files) {
-        if (file.isFile() && file.getName().endsWith(".jar")) {
-          jars.add(file);
-        }
-      }
-    }
-    return jars.toArray(new File[jars.size()]);
-  }
-
-  // TODO(joaomartins): This method should probably be extracted to CloudSdkAppEngineHelper.
-  @VisibleForTesting
-  CloudSdk buildCloudSdkWithPath(@NotNull Path path) {
-    return new CloudSdk.Builder().sdkPath(path).build();
   }
 }
